+++ playbacktests/src/androidTest/java/com/google/android/exoplayer2/playbacktests/gts/CommonEncryptionDrmTest.java
@@ -75,20 +75,6 @@ public final class CommonEncryptionDrmTest {
   }
 
   @Test
-  public void cbc1SchemeTypeV25() {
-    if (Util.SDK_INT < 25) {
-      // cbc1 support was added in API 24, but it is stable from API 25 onwards.
-      // See [internal: b/65634809].
-      // Pass.
-      return;
-    }
-    testRunner
-        .setStreamName("test_widevine_h264_scheme_cbc1")
-        .setManifestUrl(DashTestData.WIDEVINE_SCHEME_CBC1)
-        .run();
-  }
-
-  @Test
   public void cbcsSchemeTypeV25() {
     if (Util.SDK_INT < 25) {
       // cbcs support was added in API 24, but it is stable from API 25 onwards.
@@ -100,10 +86,5 @@ public final class CommonEncryptionDrmTest {
         .setStreamName("test_widevine_h264_scheme_cbcs")
         .setManifestUrl(DashTestData.WIDEVINE_SCHEME_CBCS)
         .run();
-  }
-
-  @Test
-  public void censSchemeTypeV25() {
-    // TODO: Implement once content is available. Track [internal: b/31219813].
   }
 }